+++ lib/src/animation_player.dart
@@ -13,12 +13,22 @@ import 'package:meta/meta.dart';
 ///
 /// To use [AnimationPlayer], provide a [PlayableAnimation] by creating one with
 /// a series of [Phase]s.
+///
+/// Your animation might also involve user interaction, in which case both this
+/// AnimationPlayer and the user's touch events need to be able to move the
+/// animation forward and backward. To accomplish this dual control, provide a
+/// [PhaseController] to this [AnimationPlayer] and then connect the user's
+/// touch events to the [PhaseController]. This way your animation will move
+/// when the user interacts with it AND when you press 'prev' and 'next in this
+/// AnimationPlayer.
 class AnimationPlayer extends StatefulWidget {
 
   final PlayableAnimation playableAnimation;
+  final PhaseController phaseController;
 
   AnimationPlayer({
     @required this.playableAnimation,
+    this.phaseController,
   }) {
     assert(playableAnimation != null);
   }
@@ -31,75 +41,102 @@ class _MultiPhaseAnimationTesterState extends State<AnimationPlayer> with Ticker
 
   static const STANDARD_PHASE_TIME = 250;
 
-  int activePhase = 0;
-  double phaseProgress = 0.0;
+  PhaseController phaseController;
   double playbackSpeed = 1.0;
-  bool playingForward = true;
   AnimationController animationController;
 
   @override
   void initState() {
     super.initState();
 
+    phaseController = widget.phaseController ?? new PhaseController();
+
     animationController = new AnimationController(
       vsync: this,
     )
       ..addListener(() {
         setState(() {
-          phaseProgress = playingForward
+          final phaseProgress = phaseController.playingForward
               ? animationController.value
               : 1.0 - animationController.value;
-
-          final animationPhase = widget.playableAnimation.phases[activePhase];
-          if (playingForward) {
-            animationPhase.forward(phaseProgress);
-          } else {
-            if (animationPhase.isUniform) {
-              animationPhase.reverse(1.0 - phaseProgress);
-            } else {
-              animationPhase.reverse(phaseProgress);
-            }
-          }
+          phaseController.update(phaseProgress: phaseProgress);
         });
       })
       ..addStatusListener((status) {
         if (AnimationStatus.forward == status) {
-          setState(() => playingForward = true);
+          setState(() => phaseController.update(playingForward: true));
         } else if (AnimationStatus.reverse == status) {
-          setState(() => playingForward = false);
+          setState(() => phaseController.update(playingForward: false));
         } else if (AnimationStatus.dismissed == status) {
-          print('Animation reached 0.0.');
+          // no-op
         } else if (AnimationStatus.completed == status) {
-          print('Animation reached 1.0.');
           // If we just finished playing forward, increment active phase.
-          if (playingForward && activePhase < widget.playableAnimation.phases.length) {
-            ++activePhase;
-            phaseProgress = 0.0;
+          if (phaseController.playingForward) {
+            if (phaseController.activePhase < widget.playableAnimation.phases.length - 1) {
+              phaseController.update(
+                activePhase: phaseController.activePhase + 1,
+                phaseProgress: 0.0,
+              );
+            } else if (phaseController.phaseProgress < 1.0) {
+              phaseController.update(
+                phaseProgress: 1.0,
+              );
+            }
           }
         }
-
-        print('New activePhase: $activePhase, phaseProgress: $phaseProgress');
       });
+
+    phaseController.addListener(_onPhaseChange);
   }
 
   @override
   void dispose() {
+    phaseController.removeListener(_onPhaseChange);
     animationController.dispose();
     super.dispose();
   }
 
+  _onPhaseChange() {
+    setState(() {
+      final animationPhase = widget.playableAnimation.phases[phaseController.activePhase];
+      if (phaseController.playingForward) {
+        animationPhase.forward(phaseController.phaseProgress);
+      } else {
+        if (animationPhase.isUniform) {
+          animationPhase.reverse(1.0 - phaseController.phaseProgress);
+        } else {
+          animationPhase.reverse(phaseController.phaseProgress);
+        }
+      }
+    });
+  }
+
   _playPhaseForward() {
-    if (activePhase < widget.playableAnimation.phases.length) {
+    if (phaseController.activePhase < widget.playableAnimation.phases.length - 1
+        || phaseController.phaseProgress < 1.0) {
       animationController.duration = new Duration(milliseconds: (STANDARD_PHASE_TIME / playbackSpeed).round());
       animationController.forward(from: 0.0);
     }
   }
 
   _playPreviousPhaseInReverse() {
-    if (activePhase >= 1) {
+    if (phaseController.activePhase >= 1 || phaseController.phaseProgress < 1.0) {
       setState(() {
-        --activePhase;
-        phaseProgress = 1.0;
+        if ((!phaseController.playingForward && phaseController.phaseProgress < 1.0)
+            || (phaseController.playingForward && phaseController.phaseProgress > 0.0)) {
+          // Take the phase progress all the way down.
+          phaseController.update(
+            phaseProgress: 1.0,
+            playingForward: false,
+          );
+        } else {
+          // Decrement the entire phase.
+          phaseController.update(
+            activePhase: phaseController.activePhase - 1,
+            phaseProgress: 0.0,
+            playingForward: false,
+          );
+        }
 
         animationController.duration = new Duration(milliseconds: (STANDARD_PHASE_TIME / playbackSpeed).round());
         animationController.reverse(from: 1.0);
@@ -111,17 +148,15 @@ class _MultiPhaseAnimationTesterState extends State<AnimationPlayer> with Ticker
     List<Widget> phases = [];
     for (var i = 0; i < widget.playableAnimation.phases.length; ++i) {
       var percent = 0.0;
-      final isActivePhase = i == activePhase;
+      final isActivePhase = i == phaseController.activePhase;
 
-      if (i < activePhase) {
+      if (i < phaseController.activePhase) {
         percent = 1.0;
-      } else if (playingForward && isActivePhase) {
-        percent = phaseProgress;
-      } else if (!playingForward && isActivePhase) {
-        percent = 1.0 - phaseProgress;
+      } else if (phaseController.playingForward && isActivePhase) {
+        percent = phaseController.phaseProgress;
+      } else if (!phaseController.playingForward && isActivePhase) {
+        percent = 1.0 - phaseController.phaseProgress;
       }
-
-      print("Playback percent: $percent");
 
       phases.add(
           new Expanded(
@@ -228,6 +263,83 @@ class PlayableAnimation {
   PlayableAnimation({
     @required this.phases,
   });
+}
+
+/// Use a PhaseController to instruct the AnimationPlayer to advance its
+/// playhead to the given Phase and percent.
+///
+/// A PhaseController provides the inverse control to the AnimationPlayer.
+/// Normally, the AnimationPlayer progresses through the Phases based on a
+/// running animation. But sometimes your animation is tied to user input
+/// so if the user taps on your UI and causes the UI to change independent
+/// of the AnimationPlayer, you need to use a PhaseController to synchronize
+/// the animation progress with the AnimationPlayer.
+class PhaseController extends ChangeNotifier {
+
+  final phaseCount;
+  int _activePhase = 0;
+  double _phaseProgress = 0.0;
+  bool _playingForward = true;
+
+  PhaseController({
+    this.phaseCount = 1,
+  }) {
+    assert(phaseCount > 0, "Can't control zero animation phases.");
+  }
+
+  get activePhase {
+    return _activePhase;
+  }
+
+  get phaseProgress {
+    return _phaseProgress;
+  }
+
+  get playingForward {
+    return _playingForward;
+  }
+
+  nextPhase() {
+    final phasePosition = _phasePosition();
+
+    if (phasePosition < phaseCount) {
+      _activePhase++;
+    } else {
+      _activePhase = phaseCount - 1;
+    }
+    _phaseProgress = 0.0;
+    _playingForward = true;
+
+    notifyListeners();
+  }
+
+  prevPhase() {
+    final phasePosition = _phasePosition();
+
+    if (phasePosition >= 1) {
+      _activePhase--;
+    } else {
+      _activePhase = 0;
+    }
+    _phaseProgress = 0.0;
+    _playingForward = true;
+
+    notifyListeners();
+  }
+
+  _phasePosition() {
+    final progressVector = playingForward ? phaseProgress : -phaseProgress;
+    return activePhase + progressVector;
+  }
+
+  update({int activePhase, double phaseProgress, playingForward}) {
+    _activePhase = activePhase ?? _activePhase;
+    _phaseProgress = phaseProgress ?? _phaseProgress;
+    _playingForward = playingForward ?? _playingForward;
+
+    notifyListeners();
+  }
+
 }
 
 /// A transition that supports forward and reverse playback.